Added season service test with some additional changes
I added a new property to the season, because previously
there was missing the season name i.e. "Season 1". Also
added a method to the season service that allows to add
a collection of seasons, beside of that I wrote a
simple test that checks the functionality of season service.